Add reservation model
On this step we introduce the simple Reservation model
with all the required fields and validations to make sure
we store valid data.

Note how we store price and currency. Price is always are integer and
it represents the smallest amount if currency ex. cent. so the
{ price: 1000, currency: 'usd'} actually means "$10.00"
This can be parsed and processed via 'money' gem that is widely used
in Ruby world.

I intentionally skip the testing on this step, since we building the
API it makes much more sence to inroduce the integration tests that
will cover all the client requirements does not matter about
the backend implementation.