Add integration tests
Let's write some integration tests that would check
the actual API request/response cycle.

I highly recomment to use Rspec for testing in Rails applications,
but since the current one is in fact one resource test application
we can cover all the API with one integration test and it is fine
to use build-in rails testing system.

Here we cover all the possible scenarios with the positive and negative
resposes from the server. Let's implement this functionality
in the following commits.