Reservations controller implementation
Here is the simple implemantation that makes our test green.
Now we can manipulate objects via the API.

Since JSON was choosen as a default format we can use any way
of formatting our objects and responses like I did in in the current
project, but in the real life projects that can have nore than
one client (web and mobile for example) I would recommend
to choose some existing standards, for example we can try to follow
JSON::API https://jsonapi.org/ but most of the existing
ruby implementations relies on ActiveRecord, or we can even try
to implement our API as a GraphQL https://graphql.org/

The current implementation can be improved in the following commits,
for eample one of the obvious changes is the pagination for index page.
But I believe that for the test assesment the current implementation
is enought.